requirements.txt
requests
beautifulsoup4
dspy-ai
google-genai
google-generativeai
pillow
